extracted UserSettingsClient to its own library
